Split ST0
