+++ config/splat.pspeu.stst0.yaml
@@ -43,7 +43,45 @@ segments:
     subalign: 8
     subsegments:
       - [0x80, c, st0_psp/80]
+      - [0x1E8, c, st0_psp/30030]
+      - [0x1578, c, st0_psp/cutscene]
+      - [0x3F58, c, st0_psp/2A8DC]
+      - [0x4A30, c, st0_psp/2B0C8]
+      - [0x58B8, c, st0_psp/2C564]
+      - [0x8230, c, st0_psp/2DAC8]
+      - [0xB3A0, c, st0_psp/3101C]
+      - [0xC3E0, c, st0_psp/3AB08]
+      - [0xE658, c, st0_psp/3C5C0]
+      - [0xF320, c, st0_psp/st_init_st0]
+      - [0xF560, c, st0_psp/2850C]
+      - [0xFD18, c, st0_psp/st_debug]
+      - [0xFDF8, c, st0_psp/2A218]
+      - [0x10678, c, st0_psp/e_lock_camera]
+      - [0x10A00, c, st0_psp/create_entity]
+      - [0x119A8, c, st0_psp/e_red_door]
+      - [0x12CB0, c, st0_psp/st_update]
+      - [0x131E8, c, st0_psp/e_particles_maria1]
+      - [0x13770, c, st0_psp/e_particles_maria2]
+      - [0x13C00, c, st0_psp/e_particles]
+      - [0x14160, c, st0_psp/e_collect]
+      - [0x15CC8, c, st0_psp/e_misc]
+      - [0x16278, c, st0_psp/collision]
+      - [0x18000, c, st0_psp/popup]
+      - [0x185F8, c, st0_psp/e_room_fg]
+      - [0x18710, c, st0_psp/st_common]
+      - [0x194C0, c, st0_psp/3D8F0]
+      - [0x19D90, c, st0_psp/prim_helpers]
+      - [0x1A4D0, c, st0_psp/e_bg_vortex]
       - [0x1B200, data]
-      - [0x42400, .rodata, st0_psp/80]
+      - [0x42400, .rodata, st0_psp/30030]
+      - [0x42450, .rodata, st0_psp/cutscene]
+      - [0x424F8, .rodata, st0_psp/2B0C8]
+      - [0x42530, .rodata, st0_psp/2C564]
+      - [0x42580, .rodata, st0_psp/2DAC8]
+      - [0x425D8, .rodata, st0_psp/3AB08]
+      - [0x42610, .rodata, st0_psp/2A218]
+      - [0x42650, .rodata, st0_psp/e_red_door]
+      - [0x42668, .rodata, st0_psp/e_collect]
+      - [0x426A0, .rodata, st0_psp/3D8F0]
       - {type: bss, vram: 0x9279D80}
   - [0x42700]
