+++ config/symbols.pspeu.stst0.txt
@@ -1,19 +1,36 @@
+SetGameState = 0x09237868;
+func_801B0058 = 0x092378A0;
+func_801B0180 = 0x092379F8;
+SetDisplayBufferColorsToBlack = 0x09237A40;
+func_801B01F8 = 0x09237A78;
+func_801B0280 = 0x09237B18;
+SetTitleDisplayBuffer = 0x09237BC0;
+func_801B0414 = 0x09237CB8;
+PrologueScroll = 0x09237D10;
 ST0_EntityCutscene = 0x092396A0;
 EntitySecretButton = 0x0923B5D8;
 EntitySecretStairsCeiling = 0x0923B950;
 EntitySecretStairs = 0x0923BA88;
 EntityStageTitleFadeout = 0x0923C0B0;
 EntityStageTitleCard = 0x0923C728;
+func_801ABBBC = 0x0923CF38;
 EntityDracula = 0x0923DC00;
 EntityDraculaBody = 0x0923F058;
 EntityDraculaFireball = 0x0923F190;
 EntityDraculaMeteorball = 0x0923F308;
 func_801AD838 = 0x0923F568;
 EntityDraculaGlass = 0x0923F618;
+func_801ADAC8 = 0x0923F8B0;
 EntityDraculaFinalForm = 0x0923F938;
 EntityDraculaMegaFireball = 0x09241268;
 EntityDraculaRainAttack = 0x09241448;
+func_801AF380 = 0x09241708;
+func_801AF6D0 = 0x09241BF0;
 func_801AF774 = 0x09241D60;
+func_801B101C = 0x09242A98;
+func_801B1198 = 0x09242EA0;
+func_801B11E8 = 0x09242F00;
+func_801B1298 = 0x09242FF8;
 EntityClouds = 0x09243A60;
 EntityClockTower3D = 0x092444B0;
 EntityCutscenePhotograph = 0x09244A80;
@@ -22,11 +39,15 @@ func_801BC5C0 = 0x09245CD8;
 func_801A805C = 0x09246BE0;
 func_801A8328 = 0x09246FB8;
 ST0_EntityBackgroundBlock = 0x09247398;
+func_801AA218 = 0x09247478;
 EntityCutscene = 0x09247550;
+PlayerIsWithinHitbox = 0x09247CF8;
 EntityLockCamera = 0x09247DF8;
 InitRoomEntities = 0x09248C08;
 UpdateTilemap = 0x09248E38;
+EntityIsNearPlayer = 0x09249028;
 ST0_EntityRedDoor = 0x09249110;
+Random = 0x0924A330;
 UpdateStageEntities = 0x0924A370;
 UpdateStageEntitiesAlt = 0x0924A750;
 EntitySoulStealOrb = 0x0924B280;
@@ -39,9 +60,16 @@ EntityGreyPuff = 0x0924D608;
 EntityIntenseExplosion = 0x0924D788;
 TestCollisions = 0x0924D8F8;
 EntityDamageDisplay = 0x0924EC78;
+BottomCornerText = 0x0924F680;
 EntityRoomForeground = 0x0924FC78;
+UnkCollisionFunc3 = 0x09250130;
 EntityDummy = 0x092508A8;
+CheckFieldCollision = 0x092508E0;
+ReplaceBreakableWithItemDrop = 0x09250A70;
 func_801BD8F0 = 0x09250B40;
+PrimResetNext = 0x09251910;
+UnkPolyFunc2 = 0x09251A18;
+PrimDecreaseBrightness = 0x09251A80;
 EntityBackgroundVortex = 0x09251B50;
 g_SpriteBanks = 0x092528E8;
 g_TileLayers = 0x09252998;
