+++ src/st/st0_psp/2A8DC.c
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2A8DC", EntitySecretButton);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2A8DC", EntitySecretStairsCeiling);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2A8DC", EntitySecretStairs);
