+++ src/st/st0_psp/2C564.c
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", func_801ABBBC);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", EntityDracula);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", EntityDraculaBody);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", EntityDraculaFireball);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", EntityDraculaMeteorball);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", func_801AD838);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2C564", EntityDraculaGlass);
