+++ src/st/st0_psp/2DAC8.c
@@ -1,0 +1,18 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+// func_801ADB10 not present
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", func_801ADAC8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", EntityDraculaFinalForm);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", EntityDraculaMegaFireball);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", EntityDraculaRainAttack);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", func_801AF380);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", func_801AF6D0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/2DAC8", func_801AF774);
