+++ src/st/st0_psp/30030.c
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", SetGameState);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", func_801B0058);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", func_801B0180);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", SetDisplayBufferColorsToBlack);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", func_801B01F8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", func_801B0280);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", SetTitleDisplayBuffer);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", func_801B0414);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/30030", PrologueScroll);
