+++ src/st/st0_psp/3101C.c
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_pspeu_09242A20);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_pspeu_09242A58);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_801B101C);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_801B1198);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_801B11E8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_pspeu_09242F68);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3101C", func_801B1298);
