+++ src/st/st0_psp/3AB08.c
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3AB08", EntityClouds);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3AB08", EntityClockTower3D);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3AB08", EntityCutscenePhotograph);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3AB08", EntityCutscenePhotographFire);
