+++ src/st/st0_psp/3D8F0.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../st0/st0.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/3D8F0", func_801BD8F0);
