+++ src/st/st0_psp/80.c
@@ -4,245 +4,3 @@
 INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237700);
 
 INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092377C8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237868);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092378A0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092379F8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237A40);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237A78);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237B18);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237BC0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237CB8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09237D10);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09238BF8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09238F10);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09238F90);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09239060);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092390C0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092392D0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09239308);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", ST0_EntityCutscene);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntitySecretButton);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntitySecretStairsCeiling);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntitySecretStairs);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityStageTitleFadeout);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityStageTitleCard);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0923CF38);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDracula);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaBody);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaFireball);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaMeteorball);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_801AD838);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaGlass);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0923F8B0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaFinalForm);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaMegaFireball);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDraculaRainAttack);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09241708);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09241BF0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_801AF774);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242A20);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242A58);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242A98);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242EA0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242F00);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242F68);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09242FF8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityClouds);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityClockTower3D);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityCutscenePhotograph);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityCutscenePhotographFire);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_801BC5C0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092469A0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09246B10);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09246B88);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_801A805C);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_801A8328);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", ST0_EntityBackgroundBlock);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09247478);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityCutscene);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09247CF8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityLockCamera);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248080);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092481A0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248338);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092484D0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248540);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092485C8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248708);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248860);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092488D8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248960);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248AA8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", InitRoomEntities);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", UpdateTilemap);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248F18);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09248FA0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09249028);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", ST0_EntityRedDoor);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924A330);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", UpdateStageEntities);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", UpdateStageEntitiesAlt);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924A868);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924ADF0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntitySoulStealOrb);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924B7E0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924B8B8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924BA38);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924BB08);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924BBA0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924BD58);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityPrizeDrop);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityExplosion);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityEquipItemDrop);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityUnkId13);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityExplosionVariants);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityGreyPuff);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityIntenseExplosion);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", TestCollisions);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDamageDisplay);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924F680);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityRoomForeground);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924FD90);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924FE28);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_0924FEB0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250020);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092500A8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092500F8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250130);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092503E8);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250458);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250500);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250580);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250670);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092506C0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250700);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityDummy);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_092508E0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09250A70);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_801BD8F0);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09251410);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09251910);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09251A18);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", func_pspeu_09251A80);
-
-INCLUDE_ASM("st/st0_psp/psp/st0_psp/80", EntityBackgroundVortex);
