+++ src/st/st0_psp/collision.c
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/collision", TestCollisions);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/collision", EntityDamageDisplay);
