+++ src/st/st0_psp/create_entity.c
@@ -1,0 +1,32 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248080);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_092481A0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248338);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_092484D0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248540);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_092485C8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248708);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248860);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_092488D8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248960);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248AA8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", InitRoomEntities);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", UpdateTilemap);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248F18);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/create_entity", func_pspeu_09248FA0);
