+++ src/st/st0_psp/cutscene.c
@@ -1,0 +1,18 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_09238BF8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_09238F10);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_09238F90);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_09239060);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_092390C0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_092392D0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", func_pspeu_09239308);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/cutscene", ST0_EntityCutscene);
