+++ src/st/st0_psp/e_collect.c
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", func_pspeu_0924B7E0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", func_pspeu_0924B8B8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", func_pspeu_0924BA38);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", func_pspeu_0924BB08);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", func_pspeu_0924BBA0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", func_pspeu_0924BD58);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", EntityPrizeDrop);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", EntityExplosion);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_collect", EntityEquipItemDrop);
