+++ src/st/st0_psp/e_misc.c
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_misc", EntityUnkId13);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_misc", EntityExplosionVariants);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_misc", EntityGreyPuff);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_misc", EntityIntenseExplosion);
