+++ src/st/st0_psp/e_particles_maria2.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_particles_maria2", func_pspeu_0924ADF0);
