+++ src/st/st0_psp/e_red_door.c
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_red_door", EntityIsNearPlayer);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_red_door", ST0_EntityRedDoor);
