+++ src/st/st0_psp/e_room_fg.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/e_room_fg", EntityRoomForeground);
