+++ src/st/st0_psp/prim_helpers.c
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../st0/st0.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/prim_helpers", func_pspeu_09251410);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/prim_helpers", PrimResetNext);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/prim_helpers", UnkPolyFunc2);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/prim_helpers", PrimDecreaseBrightness);
