+++ src/st/st0_psp/st_common.c
@@ -1,0 +1,36 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_0924FD90);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_0924FE28);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_0924FEB0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_09250020);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_092500A8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_092500F8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", UnkCollisionFunc3);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_092503E8);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_09250458);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_09250500);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_09250580);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_09250670);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_092506C0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", func_pspeu_09250700);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", EntityDummy);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", CheckFieldCollision);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_common", ReplaceBreakableWithItemDrop);
