+++ src/st/st0_psp/st_debug.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_debug", ST0_EntityBackgroundBlock);
