+++ src/st/st0_psp/st_init_st0.c
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_init_st0", func_pspeu_092469A0);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_init_st0", func_pspeu_09246B10);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_init_st0", func_pspeu_09246B88);
