+++ src/st/st0_psp/st_update.c
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_update", Random);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_update", UpdateStageEntities);
+
+INCLUDE_ASM("st/st0_psp/psp/st0_psp/st_update", UpdateStageEntitiesAlt);
